# perf: outline logic in Decode to allow for stack allocations