--- a/version.json
+++ b/version.json
@@ -1,3 +1,3 @@
 {
-	"version": "v0.2.2"
+	"version": "v0.2.3"
 }